+++ pot-bot-app/src/presenters/Plant.js
@@ -1,10 +1,8 @@
 import React, {useEffect, useState} from "react";
-import {Link, useNavigate} from "react-router-dom";
 import elephant from "../styling/images/elefant.jpg";
 import {useAuth} from "../firebaseModel";
-import addPlantIcon from '../styling/images/plus-pot.png'
 
-export function Plants({plant, data, today}) {
+export function Plant({plant, data, today}) {
   const [expanded, setExpanded] = useState(false);
   const [latest, setLatest] = useState(null)
   const [timeIndex, setTimeIndex] = useState("")
@@ -18,7 +16,7 @@ export function Plants({plant, data, today}) {
     let latestDate = today
     //const d = new Date()
     let index = timeIndex
-    if(timeIndex === "" && data !== null) {
+    if (timeIndex === "" && data !== null) {
       //let time = (`${d.getHours()}:${d.getMinutes()}`)
       Object.keys(data).forEach(key => {
         if (today !== key) {
@@ -27,18 +25,18 @@ export function Plants({plant, data, today}) {
       })
       index = Object.entries(data[latestDate]).sort().reverse().at(0).at(0)
     }
-      setLatest(data[latestDate][index])
-      setTimeIndex(index)
+    setLatest(data[latestDate][index])
+    setTimeIndex(index)
   }, [user, data])
   console.log(latest)
-  return(
+  return (
     <>
-    <div className={`expandable-div ${expanded ? "expanded" : ""}`}
-         onClick={handleClick}>
-      <div className="card-title">
-        <img src={elephant} width="100" height="100"/>
-        <span style={{fontFamily: "sans-serif", padding: "0.5em"}}>{plant}</span>
-      </div>
+      <div className={`expandable-div ${expanded ? "expanded" : ""}`}
+           onClick={handleClick}>
+        <div className="card-title">
+          <img src={elephant} width="100" height="100"/>
+          <span style={{fontFamily: "sans-serif", padding: "0.5em"}}>{plant}</span>
+        </div>
         {expanded && <div className="plant-data">
           <div>Moisture: {latest.soilMoisture} </div>
           <div>Light: {latest.uvIntensity}</div>
@@ -46,7 +44,7 @@ export function Plants({plant, data, today}) {
           <div>Waterlevel: {latest.waterLevel}</div>
         </div>}
       </div>
-    {/* <div className={"addPlant"}>
+      {/* <div className={"addPlant"}>
       <Link to="/addNewPlant"><img src={addPlantIcon}/></Link>
     </div> */}
     </>)
