+++ pot-bot-app/src/presenters/ResetPasswordPresenter.js
@@ -1,0 +1,11 @@
+
+import React from 'react';
+import ResetPasswordView from "../views/ResetPasswordView";
+
+function ResetPasswordPresenter() {
+    return (
+        <ResetPasswordView/>
+    );
+}
+
+export default ResetPasswordPresenter;
