--- pot-bot-app/src/presenters/resetPasswordPresenter.js
@@ -1,11 +1,0 @@
-
-import React from 'react';
-import ResetPasswordView from "../views/resetPasswordView";
-
-function ResetPresenter() {
-    return (
-        <ResetPasswordView/>
-    );
-}
-
-export default ResetPresenter;
