+++ pot-bot-app/src/views/ResetPasswordView.js
